+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.38.5
 
+* Enhancement: Add OS condition #989
 * Enhancement: Enable setting default profile by env variable #996 (thanks @Johnabell)
 * Documentation: Add Fig Autocompletion to README.md #993 (thanks @beeinger)
 * Maintenance: update dependencies and build